core: PresetMetadata → EffectDef converter + dual-source registry
Block 2 of the §11 unified-registry migration. Two pieces:

1. `preset_metadata_to_effect_def(&PresetMetadata) -> EffectDef`
   — converts a parsed JSON preset to the existing EffectDef shape
   that 50+ consumers in effects.rs and project.rs already walk.
   Owned Strings move to &'static str via Box::leak (bounded by
   preset count, startup-only — same lifetime as inventory data,
   just sourced differently).

2. `loaded_preset_metadata() -> &'static [PresetMetadata]` — stable
   API stub for the JSON-loaded preset slice. Returns empty today;
   block 3 (build.rs codegen) and block 4 (per-effect migration)
   populate it. The `DEFINITIONS` LazyLock now iterates this slice
   alongside the existing inventory submissions, with JSON-loaded
   entries winning over inventory for the same id. No observable
   behavior change yet (empty slice = no overrides).

Tests confirm the new path produces an EffectDef equivalent to the
inventory path for the same effect shape — the parity invariant
that block 4's per-effect migration relies on.

Drive-by: `ValueAliasEntry.mapping` is `Vec<(i32, i32)>` not
`Vec<(f32, f32)>` to match the renderer-side
`ParamValueAlias = (i32, i32)` shape for enum-value remaps.

`BindingDef` and `SkipModeDef` are defined but not yet projected
into EffectDef — they're renderer-side concerns and will land on a
renderer-side `LoadedPreset` view in block 6.

+++ crates/manifold-core/src/effect_graph_def.rs
@@ -298,6 +298,9 @@ pub struct AliasEntry {
 /// One entry in a value-remap alias table — applies to a single
 /// param's stored value at load time. Used when an effect's enum
 /// value indices shift across a refactor.
+///
+/// Matches the renderer-side `ParamValueAlias = (i32, i32)` shape
+/// for `(from, to)` pairs.
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ValueAliasEntry {
@@ -305,7 +308,7 @@ pub struct ValueAliasEntry {
     pub param_id: String,
     /// Pairs of `(stored_value, new_value)` — when the loader sees a
     /// param value matching the first, it rewrites to the second.
-    pub mapping: Vec<(f32, f32)>,
+    pub mapping: Vec<(i32, i32)>,
 }
 
 #[cfg(test)]
